Come up with a catchy name
"Go with RSpec" is a nice name if all you ever wanted to do was test Go.
But what about Haskell? What about your command-line app written in java
that is compiled to Dart that is then converted to Javascript?
The possibilities are endless here. We need a name that recognizes the
enormous potential of this concept.
Test Go, sure. But you can you also test an enterprise C# application.
We're going off the rails here. Test anything. Test Lisp. Test Rust.
Test a microwave. Your toaster. Test literally anything.